provide `.parentExpression` for expressions only
* re https://shiftleftsecurity.slack.com/archives/C3TDFSH8X/p1587636696453700?thread_ts=1587632856.441400&cid=C3TDFSH8X
* tested with codescience build
