+++ semanticcpg/src/main/scala/io/shiftleft/semanticcpg/language/nodemethods/AstNodeMethods.scala
@@ -1,12 +1,8 @@
 package io.shiftleft.semanticcpg.language.nodemethods
 
-import io.shiftleft.codepropertygraph.generated.nodes
 import io.shiftleft.Implicits.JavaIteratorDeco
-import io.shiftleft.semanticcpg.language.types.expressions.generalizations.AstNode
+import io.shiftleft.codepropertygraph.generated.nodes
 import io.shiftleft.semanticcpg.language._
-import io.shiftleft.semanticcpg.utils.MemberAccess
-
-import scala.annotation.tailrec
 
 class AstNodeMethods(val node: nodes.AstNode) extends AnyVal {
 
@@ -58,24 +54,6 @@ class AstNodeMethods(val node: nodes.AstNode) extends AnyVal {
                        } else {
                          childDepths.max
                        })
-  }
-
-  /**
-    * Traverse to it's parent expression (e.g. call or return) by following the incoming AST nodes.
-    * It's continuing it's walk until it hits an expression that's not a generic
-    * "member access operation", e.g., "<operator>.memberAccess".
-    * */
-  def parentExpression: nodes.Expression = _parentExpression(node)
-
-  @tailrec
-  final def _parentExpression(argument: nodes.AstNode): nodes.Expression = {
-    val parent = argument._astIn.onlyChecked
-    parent match {
-      case call: nodes.Call if MemberAccess.isGenericMemberAccessName(call.name) =>
-        _parentExpression(call)
-      case expression: nodes.Expression =>
-        expression
-    }
   }
 
   def astParent: nodes.AstNode =
